fix(classification_functions): correct dtype annotations for class IDs
- lin_class: class_id parameter is int64 (matches docstring); first return
  element (class array) is int64.
- mahal_class: first return element (mahal_class_arr) is int64.
- norm_class: first return element (norm_class_id) is int64.
- gen_two_step_class_stats: mahal_class_id return is int64.

Closes #148

+++ src/rock_physics_open/equinor_utilities/classification_functions/mahal_class.py
@@ -12,7 +12,7 @@ def mahal_class(
     class_cov: npt.NDArray[np.float64],
     class_id: npt.NDArray[np.int64],
     thresh: float = np.inf,
-) -> tuple[npt.NDArray[np.float64], npt.NDArray[np.float64], npt.NDArray[np.float64]]:
+) -> tuple[npt.NDArray[np.int64], npt.NDArray[np.float64], npt.NDArray[np.float64]]:
     """
     Mahalanobis classification routine. All data points are assigned a class, unless a threshold is set.
